# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Bancaire",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.4.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/ly/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/comptes/{numero}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer un compte par numéro",
                "description": "Retourne les détails d'un compte spécifique basé sur son numéro",
                "operationId": "getCompteByNumero",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte trouvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte trouvé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CompteResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comptes/client/{telephone}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer les comptes d'un client par téléphone",
                "description": "Retourne la liste des comptes d'un client basé sur son numéro de téléphone",
                "operationId": "getComptesByTelephone",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "path",
                        "description": "Numéro de téléphone du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comptes trouvés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes trouvés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CompteResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun compte trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun compte trouvé pour ce numéro de téléphone"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CompteResource": {
                "title": "Compte Resource",
                "description": "Représentation d'un compte bancaire",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du compte",
                        "type": "string",
                        "format": "uuid"
                    },
                    "numeroCompte": {
                        "description": "Numéro du compte",
                        "type": "string"
                    },
                    "titulaire": {
                        "description": "Nom du titulaire du compte",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "epargne",
                            "courant",
                            "cheque"
                        ]
                    },
                    "solde": {
                        "description": "Solde du compte",
                        "type": "number",
                        "format": "float"
                    },
                    "devise": {
                        "description": "Devise du compte",
                        "type": "string"
                    },
                    "dateCreation": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date-time"
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque"
                        ]
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "description": "Dernière modification",
                                "type": "string",
                                "format": "date-time"
                            },
                            "version": {
                                "description": "Version de l'API",
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}